+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+* Fix bug in `ContextIO::Message#set_flags` that was failing, and also fixes the wrong api url it was trying to call. - [Paul Bonaud](https://github.com/popox)
+
 ## 1.7.0
 
 * Further README updates for clarity. - Johnny Goodman